back/requirements.txt
mysql-connector-python
PyJWT
bcrypt